% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manBayesMedILDPrev-package.R
\docType{package}
\name{manBayesMedILDPrev-package}
\alias{manBayesMedILDPrev-package}
\alias{_PACKAGE}
\title{manBayesMedILDPrev: Bayesian Mediation Modeling for Intensive Longitudinal Data in Prevention Science}
\description{
Research compendium for the manuscript Melon, B., et al. (2023). Bayesian Mediation Modeling for Intensive Longitudinal Data in Prevention Science. \doi{10.0000/0000000000}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sigmaresearch100/manBayesMedILDPrev}
  \item \url{https://sigmaresearch100.github.io/manBayesMedILDPrev/}
  \item \url{https://osf.io/zae7s/?view_only=9d2d5366d7624f95beca835cd7315c38}
  \item \doi{10.0000/0000000000}
  \item Report bugs at \url{https://github.com/sigmaresearch100/manBayesMedILDPrev/issues}
}

}
\author{
\strong{Maintainer}: Blind Melon \email{r.sigmaresearch100@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID}) [copyright holder]

}
\keyword{internal}
